RBF refactoring
