+++ pygem/__init__.py
@@ -23,6 +23,8 @@ __status__ = "Stable"
 #from .affine import *
 from .deformation import Deformation
 from .ffd import FFD
+from .rbf import RBF
+from .rbf_factory import RBFFactory
 #from .radial import RBF
 #from .idw import IDW
 #from .filehandler import FileHandler
